+++ src/test/java/edu/illinois/library/cantaloupe/cache/CacheFacadeTest.java
@@ -392,6 +392,34 @@ public class CacheFacadeTest extends BaseTest {
         }
     }
 
+    /* purgeInfos() */
+
+    @Test
+    void testPurgeInfos() throws Exception {
+        final Configuration config = Configuration.getInstance();
+        config.setProperty(Key.DERIVATIVE_CACHE_TTL, 1);
+
+        enableDerivativeCache();
+        DerivativeCache derivCache = CacheFactory.getDerivativeCache().get();
+
+        Identifier identifier = new Identifier("jpg");
+        Info info             = new Info();
+
+        // Add info to the derivative cache.
+        derivCache.put(identifier, info);
+
+        // Assert that everything has been added.
+        assertNotNull(derivCache.getInfo(identifier));
+
+        instance.purgeInfos();
+
+        Thread.sleep(ASYNC_WAIT);
+
+        // Assert that it's gone.
+        assertEquals(0, InfoService.getInstance().getInfoCache().size());
+        assertFalse(derivCache.getInfo(identifier).isPresent());
+    }
+
     /* purgeInvalid() */
 
     @Test